+++ client/test/help/0/1.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+./sc-client help infect > output.txt
+diff -Z output.txt client/test/help/0/1.stdout