+++ client/test/help/0/6.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+./sc-client help unhide > output.txt
+diff -Z output.txt client/test/help/0/6.stdout