+++ client/test/help/0/9.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+./sc-client help payload > output.txt
+diff -Z output.txt client/test/help/0/9.stdout